third_party/blink/web_tests/platform/win/external/wpt/css/cssom/font-family-serialization-001-expected.txt

This is a testharness.js-based test.
[FAIL] Serialization of quoted "<generic-family>"
  assert_equals: expected "\\"emoji\\"" but got "emoji"
Harness: the test ran to completion.
